refactor(design-tokens): move Localizer into Admin\Feed namespace
